backendmodule/requirements.txt
flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
flask_sqlalchemy==3.1.1
SQLAlchemy==2.0.31
Flask-Migrate==4.0.7
Flask-Login==0.6.3
psycopg2==2.9.9
Flask-RESTful==0.3.10